:recycle: refactor: 불필요한 코드, 파일 삭제
-불필요한 코드, 파일 삭제
