+++ src/main/java/com/sparta/oishitable/domain/follow/repository/FollowRepository.java
@@ -6,7 +6,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
@@ -16,22 +15,10 @@ public interface FollowRepository extends JpaRepository<Follow, Long> {
     boolean existsByFollowerIdAndFollowingId(Long followerId, Long followingId);
 
     @Query("SELECT f.follower FROM Follow f WHERE f.following.id = :userId")
-    Page<User> findFollowersByUserId(@Param("userId") Long userId, Pageable pageable);
+    Page<User> findFollowersByFollowingId(Long userId, Pageable pageable);
 
     @Query("SELECT f.following FROM Follow f WHERE f.follower.id = :userId")
-    Page<User> findFollowingsByUserId(@Param("userId") Long userId, Pageable pageable);
+    Page<User> findFollowingsByFollowerId(Long userId, Pageable pageable);
 
-    @Query("SELECT COUNT(f) FROM Follow f WHERE f.following.id = :userId")
-    Long countFollowersByUserId(@Param("userId") Long userId);
-
-    @Query("SELECT COUNT(f) FROM Follow f WHERE f.follower.id = :userId")
-    Long countFollowingsByUserId(@Param("userId") Long userId);
-
-    void deleteByFollowerIdAndFollowingId(Long followerId, Long followingId);
-
-    @Query("SELECT f FROM Follow f WHERE f.follower.id = :followerId AND f.following.id = :followingId")
-    Optional<Follow> findByFollowerIdAndFollowingId(@Param("followerId") Long followerId, @Param("followingId") Long followingId);
-
-    @Query("SELECT CASE WHEN COUNT(f) > 0 THEN true ELSE false END FROM Follow f WHERE f.follower.id = :followerId AND f.following.id = :followingId")
-    boolean isFollowing(@Param("followerId") Long followerId, @Param("followingId") Long followingId);
+    Optional<Follow> findByFollowerIdAndFollowingId(Long followerId, Long followingId);
 }
